LayoutTests/platform/chromium-mac/fast/canvas/webgl/null-object-behaviour-expected.txt

Tests calling WebGL APIs without providing the necessary objects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.compileShader() threw exception GL error 1282 in compileShader.
PASS context.linkProgram() threw exception GL error 1282 in linkProgram.
PASS context.attachShader() threw exception GL error 1281 in attachShader.
PASS context.attachShader(program, undefined) threw exception GL error 1281 in attachShader.
PASS context.attachShader(undefined, shader) threw exception GL error 1281 in attachShader.
PASS context.detachShader(program, undefined) threw exception GL error 1282 in detachShader.
PASS context.detachShader(undefined, shader) threw exception GL error 1282 in detachShader.
PASS context.shaderSource() threw exception GL error 1282 in shaderSource.
PASS context.shaderSource(undefined, 'foo') threw exception GL error 1282 in shaderSource.
PASS context.bindAttribLocation(undefined, 0, 'foo') threw exception GL error 1282 in bindAttribLocation.
PASS context.bindBuffer(context.ARRAY_BUFFER, 0) is undefined.
PASS context.bindFramebuffer(context.FRAMEBUFFER, 0) is undefined.
PASS context.bindRenderbuffer(context.RENDERBUFFER, 0) is undefined.
PASS context.bindTexture(context.TEXTURE_2D, 0) is undefined.
PASS context.framebufferRenderbuffer(context.FRAMEBUFFER, context.DEPTH_ATTACHMENT, context.RENDERBUFFER, 0) threw exception GL error 1282 in framebufferRenderbuffer.
PASS context.framebufferTexture2D(context.FRAMEBUFFER, context.COLOR_ATTACHMENT0, context.TEXTURE_2D, 0, 0) threw exception GL error 1282 in framebufferTexture2D.
PASS context.getProgramParameter(undefined, 0) threw exception GL error 1282 in getProgramParameter.
PASS context.getProgramInfoLog(undefined, 0) threw exception GL error 1282 in getProgramInfoLog.
PASS context.getShaderParameter(undefined, 0) threw exception GL error 1282 in getShaderParameter.
PASS context.getShaderInfoLog(undefined, 0) threw exception GL error 1282 in getShaderInfoLog.
PASS context.getShaderSource(undefined) threw exception GL error 1282 in getShaderSource.
PASS context.getUniform(undefined, 0) threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS context.getUniformLocation(undefined, 'foo') threw exception GL error 1282 in getUniformLocation.
PASS successfullyParsed is true

TEST COMPLETE
